requirements.txt
python-dotenv
pandas
numpy
pymongo
pymongo[srv]
certifi
python-dotenv
scikit-learn
dill
mlflow
pyaml